chore(compose): move the two adapters out of server.go, under the size cap
internal/compose/server.go was 505 lines against a 500-line cap, so
deterministic-gates failed on a clean checkout of main and on every
branch regardless of what it changed.

Two adapters move to serveradapters.go: signalStrength, the arch-legal
signals-to-people seam, and paramParseError, the generated-parameter
error mapping. Both are self-contained and neither is part of what the
composed Server IS — server.go answers what this installation is wired
out of, and a reader answering that does not need either body.

No behaviour change. Both moved verbatim, imports follow them.
server.go is 465 lines and make go-file-length passes.
