+++ backend/internal/compose/server.go
@@ -44,11 +44,9 @@ import (
 	"github.com/gradionhq/margince/backend/internal/platform/agentquota"
 	"github.com/gradionhq/margince/backend/internal/platform/auth"
 	"github.com/gradionhq/margince/backend/internal/platform/blobstore"
-	"github.com/gradionhq/margince/backend/internal/platform/httperr"
 	"github.com/gradionhq/margince/backend/internal/platform/httpserver"
 	"github.com/gradionhq/margince/backend/internal/platform/keyvault"
 	"github.com/gradionhq/margince/backend/internal/platform/overlaybudget"
-	"github.com/gradionhq/margince/backend/internal/shared/kernel/ids"
 )
 
 // Server satisfies crmcontracts.ServerInterface by embedding: every
@@ -462,44 +460,4 @@ func (s *Server) rebuildToolRegistry(pool *pgxpool.Pool) {
 		s.replyDrafter, s.resolveOverlayIncumbent(pool), s.send, companyEnricher{srv: s},
 		s.retrievalEmbedder,
 		agents.WithQuotaCharger(s.quotaMeter), agents.WithCostShare(s.quotaMeter))
-}
-
-// signalStrength bridges people's §4 relationship-strength computation to
-// the slice the warm room consumes (signals.StrengthSource). It carries
-// only the score and its bucket across the seam — the full explainable
-// decomposition stays with its owner. This is the arch-legal edge: signals
-// declares its own seam type, and the cross-module dependency lives here in
-// compose, never as a signals→people import.
-type signalStrength struct{ people *people.Store }
-
-func (s signalStrength) PersonStrength(ctx context.Context, personID ids.PersonID, now time.Time) (signals.RelationshipStrength, error) {
-	rs, err := s.people.PersonStrength(ctx, personID, now)
-	if err != nil {
-		return signals.RelationshipStrength{}, err
-	}
-	return signals.RelationshipStrength{Strength: rs.Strength, Bucket: rs.Bucket}, nil
-}
-
-// paramParseError maps a generated request-parameter parse failure onto
-// the same 422 validation_error shape every other bad query input uses
-// (mirrors httperr's malformed-cursor path). It names only the offending
-// parameter — never the wrapped parser text, which can carry internal
-// detail — so a bad cursor/limit/sort/UUID answers problem+json, not a
-// text/plain leak.
-func paramParseError(w http.ResponseWriter, r *http.Request, err error) {
-	param := "request"
-	switch e := err.(type) {
-	case *crmcontracts.RequiredParamError:
-		param = e.ParamName
-	case *crmcontracts.InvalidParamFormatError:
-		param = e.ParamName
-	case *crmcontracts.TooManyValuesForParamError:
-		param = e.ParamName
-	case *crmcontracts.UnmarshalingParamError:
-		param = e.ParamName
-	case *crmcontracts.UnescapedCookieParamError:
-		param = e.ParamName
-	}
-	httperr.Write(w, r, httperr.Validation(param, "invalid_parameter",
-		"parameter is missing or malformed"))
 }
