+++ backend/internal/compose/serveradapters.go
@@ -1,0 +1,61 @@
+// SPDX-License-Identifier: BUSL-1.1
+// SPDX-FileCopyrightText: 2026 Gradion
+
+package compose
+
+// The two adapters the composed Server needs but does not consist of: one
+// cross-module seam, and one error mapping. They sit here rather than in
+// server.go because that file's job is to say WHAT the installation is wired
+// out of, and a reader answering that question does not need either body.
+
+import (
+	"context"
+	"net/http"
+	"time"
+
+	crmcontracts "github.com/gradionhq/margince/backend/internal/contracts"
+	"github.com/gradionhq/margince/backend/internal/modules/people"
+	"github.com/gradionhq/margince/backend/internal/modules/signals"
+	"github.com/gradionhq/margince/backend/internal/platform/httperr"
+	"github.com/gradionhq/margince/backend/internal/shared/kernel/ids"
+)
+
+// signalStrength bridges people's §4 relationship-strength computation to
+// the slice the warm room consumes (signals.StrengthSource). It carries
+// only the score and its bucket across the seam — the full explainable
+// decomposition stays with its owner. This is the arch-legal edge: signals
+// declares its own seam type, and the cross-module dependency lives here in
+// compose, never as a signals→people import.
+type signalStrength struct{ people *people.Store }
+
+func (s signalStrength) PersonStrength(ctx context.Context, personID ids.PersonID, now time.Time) (signals.RelationshipStrength, error) {
+	rs, err := s.people.PersonStrength(ctx, personID, now)
+	if err != nil {
+		return signals.RelationshipStrength{}, err
+	}
+	return signals.RelationshipStrength{Strength: rs.Strength, Bucket: rs.Bucket}, nil
+}
+
+// paramParseError maps a generated request-parameter parse failure onto
+// the same 422 validation_error shape every other bad query input uses
+// (mirrors httperr's malformed-cursor path). It names only the offending
+// parameter — never the wrapped parser text, which can carry internal
+// detail — so a bad cursor/limit/sort/UUID answers problem+json, not a
+// text/plain leak.
+func paramParseError(w http.ResponseWriter, r *http.Request, err error) {
+	param := "request"
+	switch e := err.(type) {
+	case *crmcontracts.RequiredParamError:
+		param = e.ParamName
+	case *crmcontracts.InvalidParamFormatError:
+		param = e.ParamName
+	case *crmcontracts.TooManyValuesForParamError:
+		param = e.ParamName
+	case *crmcontracts.UnmarshalingParamError:
+		param = e.ParamName
+	case *crmcontracts.UnescapedCookieParamError:
+		param = e.ParamName
+	}
+	httperr.Write(w, r, httperr.Validation(param, "invalid_parameter",
+		"parameter is missing or malformed"))
+}
